common/lru: rename new types
