+++ common/lru/basiclru_test.go
@@ -21,7 +21,7 @@ import (
 )
 
 // keys returns all keys in the cache.
-func (c *Cache[K, V]) keys() []K {
+func (c *BasicLRU[K, V]) keys() []K {
 	keys := make([]K, 0, len(c.items))
 	for node := c.list.head; node != nil; node = node.next {
 		keys = append(keys, node.v)
@@ -30,7 +30,7 @@ func (c *Cache[K, V]) keys() []K {
 }
 
 func TestCache(t *testing.T) {
-	cache := NewCache[int, int](128)
+	cache := New[int, int](128)
 
 	for i := 0; i < 256; i++ {
 		cache.Add(i, i)
@@ -78,7 +78,7 @@ func TestCache(t *testing.T) {
 
 // This test checks GetOldest and RemoveOldest.
 func TestCacheGetOldest(t *testing.T) {
-	cache := NewCache[int, int](128)
+	cache := New[int, int](128)
 	for i := 0; i < 256; i++ {
 		cache.Add(i, i)
 	}
@@ -110,7 +110,7 @@ func TestCacheGetOldest(t *testing.T) {
 
 // Test that Add returns true/false if an eviction occurred
 func TestCacheAddReturnValue(t *testing.T) {
-	cache := NewCache[int, int](1)
+	cache := New[int, int](1)
 	if cache.Add(1, 1) {
 		t.Errorf("first add shouldn't have evicted")
 	}
@@ -121,7 +121,7 @@ func TestCacheAddReturnValue(t *testing.T) {
 
 // This test verifies that Contains doesn't change item recency.
 func TestCacheContains(t *testing.T) {
-	cache := NewCache[int, int](2)
+	cache := New[int, int](2)
 	cache.Add(1, 1)
 	cache.Add(2, 2)
 	if !cache.Contains(1) {
