+++ common/lru/lru.go
@@ -14,193 +14,72 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
-// Package lru implements generically-typed LRU caches.
 package lru
 
-// Cache is a simple LRU cache.
-//
-// This type is not safe for concurrent use.
-// The zero value is not valid, instances must be created using NewCache.
-type Cache[K comparable, V any] struct {
-	list  dlist[K]
-	items map[K]cacheItem[K, V]
-	cap   int
+import "sync"
+
+// LRU is an LRU cache protected by a mutex.
+type LRU[K comparable, V any] struct {
+	cache BasicLRU[K, V]
+	mu    sync.Mutex
 }
 
-type cacheItem[K any, V any] struct {
-	value V
-	node  *dlistNode[K]
-}
-
-// NewCache creates a new LRU cache.
-func NewCache[K comparable, V any](capacity int) Cache[K, V] {
-	if capacity < 0 {
-		capacity = 1
-	}
-	return Cache[K, V]{
-		items: make(map[K]cacheItem[K, V]),
-		cap:   capacity,
-	}
+func NewLRU[K comparable, V any](capacity int) *LRU[K, V] {
+	return &LRU[K, V]{cache: NewBasicLRU[K, V](capacity)}
 }
 
 // Add adds a value to the cache. Returns true if an item was evicted to store the new item.
-func (c *Cache[K, V]) Add(key K, value V) (evicted bool) {
-	item, ok := c.items[key]
-	if ok {
-		// Already exists in cache.
-		item.value = value
-		c.list.moveToFront(item.node)
-		return false
-	}
+func (c *LRU[K, V]) Add(key K, value V) (evicted bool) {
+	c.mu.Lock()
+	defer c.mu.Unlock()
 
-	if c.Len() >= c.cap {
-		// Evict an item.
-		node := c.list.removeLast()
-		delete(c.items, node.v)
-		evicted = true
-	}
-
-	// Store the new item.
-	item = cacheItem[K, V]{value: value, node: c.list.push(key)}
-	c.items[key] = item
-	return evicted
+	return c.cache.Add(key, value)
 }
 
 // Contains reports whether the given key exists in the cache.
-func (c *Cache[K, V]) Contains(key K) bool {
-	_, ok := c.items[key]
-	return ok
+func (c *LRU[K, V]) Contains(key K) bool {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
+	return c.cache.Contains(key)
 }
 
 // Get retrieves a value from the cache. This marks the key as recently used.
-func (c *Cache[K, V]) Get(key K) (value V, ok bool) {
-	item, ok := c.items[key]
-	if !ok {
-		return value, false
-	}
-	c.list.moveToFront(item.node)
-	return item.value, true
-}
+func (c *LRU[K, V]) Get(key K) (value V, ok bool) {
+	c.mu.Lock()
+	defer c.mu.Unlock()
 
-// GetOldest retrieves the least-recently-used item.
-// Note that this does not update the item's recency.
-func (c *Cache[K, V]) GetOldest() (key K, value V, ok bool) {
-	if c.list.tail == nil {
-		return key, value, false
-	}
-	key = c.list.tail.v
-	item := c.items[key]
-	return key, item.value, true
+	return c.cache.Get(key)
 }
 
 // Len returns the current number of items in the cache.
-func (c *Cache[K, V]) Len() int {
-	return len(c.items)
+func (c *LRU[K, V]) Len() int {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
+	return c.cache.Len()
 }
 
 // Peek retrieves a value from the cache, but does not mark the key as recently used.
-func (c *Cache[K, V]) Peek(key K) (value V, ok bool) {
-	item, ok := c.items[key]
-	if !ok {
-		return value, false
-	}
-	return item.value, true
+func (c *LRU[K, V]) Peek(key K) (value V, ok bool) {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
+	return c.cache.Peek(key)
 }
 
 // Purge empties the cache.
-func (c *Cache[K, V]) Purge() {
-	c.list.init()
-	for k := range c.items {
-		delete(c.items, k)
-	}
+func (c *LRU[K, V]) Purge() {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
+	c.cache.Purge()
 }
 
 // Remove drops an item from the cache. Returns true if the key was present in cache.
-func (c *Cache[K, V]) Remove(key K) bool {
-	item, ok := c.items[key]
-	if ok {
-		delete(c.items, key)
-		c.list.remove(item.node)
-	}
-	return ok
-}
+func (c *LRU[K, V]) Remove(key K) bool {
+	c.mu.Lock()
+	defer c.mu.Unlock()
 
-// RemoveOldest drops the least recently used item.
-func (c *Cache[K, V]) RemoveOldest() (key K, value V, ok bool) {
-	if c.list.tail == nil {
-		return key, value, false
-	}
-	key = c.list.tail.v
-	item := c.items[key]
-	delete(c.items, key)
-	c.list.remove(c.list.tail)
-	return key, item.value, true
-}
-
-// dlist is a doubly-linked list holding items of type T.
-type dlist[T any] struct {
-	head *dlistNode[T]
-	tail *dlistNode[T]
-}
-
-type dlistNode[T any] struct {
-	v    T
-	next *dlistNode[T]
-	prev *dlistNode[T]
-}
-
-// init reinitializes the list, making it empty.
-func (l *dlist[T]) init() {
-	l.head, l.tail = nil, nil
-}
-
-// push adds a new item to the front of the list and returns the
-func (l *dlist[T]) push(item T) *dlistNode[T] {
-	node := &dlistNode[T]{v: item}
-	l.pushNode(node)
-	return node
-}
-
-func (l *dlist[T]) pushNode(node *dlistNode[T]) {
-	if l.head == nil {
-		// List is empty, new node is head and tail.
-		l.head = node
-		l.tail = node
-	} else {
-		node.next = l.head
-		l.head.prev = node
-		l.head = node
-	}
-}
-
-// moveToFront makes 'node' the head of the list.
-func (l *dlist[T]) moveToFront(node *dlistNode[T]) {
-	l.remove(node)
-	l.pushNode(node)
-}
-
-// remove removes an element from the list.
-func (l *dlist[T]) remove(node *dlistNode[T]) {
-	if node.next != nil {
-		node.next.prev = node.prev
-	}
-	if node.prev != nil {
-		node.prev.next = node.next
-	}
-	if l.head == node {
-		l.head = node.next
-	}
-	if l.tail == node {
-		l.tail = node.prev
-	}
-	node.next, node.prev = nil, nil
-}
-
-// removeLast removes the last element of the list.
-func (l *dlist[T]) removeLast() *dlistNode[T] {
-	last := l.tail
-	if last != nil {
-		l.remove(last)
-	}
-	return last
+	return c.cache.Remove(key)
 }
